feat(core): refactor the configurations struct and update related code
