--- .go-version
@@ -1,1 +1,0 @@
-1.23.1
